Add first delete definitions
We are keeping all definitions in this file because it is a prototype.
The aim is to divide a delete ascension in steps were focus and context
are both well formed.
The idea is that a single step does not affect the Tree (focus = DUp) or
changes it (focus = DDelete): in both cases the given focus should
contain well formed trees.