Modify step_del_ts root case
we need to modify the step tree stack in order to have a good
invariant (i.e. an operation step in which everything is well
formed). The idea is to have a merged node passed through the DDelete
constructor, and update the parent node and merge it if necessary so
that we can push it forward. We will need to have a step across, and to
navigate two bits of context at the time.

In this commit we redefine the step_del_ts to deal with the root case.