Add merge cases leaving all partial trees wf
we know our focus was merged, so we insert the focus in the parent and
delete the redundant child. Then we restructure the parent with the
grandparent: if too slim and no fat neighbours we require another
merge. In this case we make a focus that is merged with a grandparent
child, neighbour of the parent.