Edit steal right and steal left
introduced a special datatype for up after stealing, because we need to
rotate the key. We do not need to pass an index in the datatype since
the first bit of the context tells us where we need to insert the key,
and we can infer if we did a steal left or a steal right (by order of keys
or by quantities of keys in the nodes before the stealing).