proceed with find proof and generate scala
the proof showed that the wellformed_fts_1 fts definition is redundant.
Also there is an error in wf_context_1 definition:
if we are in the root and we are setting the lower and upper bound keys,
the definition tries to check these keys against themselves (see
check_keys bit).

This commit enables the scala generation by omitting constants definitions.